@@ -3,12 +3,17 @@ package com.kindsonthegenius.fleetmsv2.accounts.models;
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import com.kindsonthegenius.fleetmsv2.parameters.models.Client;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
 
-import javax.persistence.*;
 import java.util.Date;
 
 @Entity
@@ -24,9 +29,9 @@ public class Invoice {
 	
 	@DateTimeFormat(pattern = "yyyy-MM-dd")
 	private Date invoiceDate;
-	
+
 	@ManyToOne
-	@JoinColumn(name="invoicestatusid", insertable=false, updatable=false)	
+	@JoinColumn(name = "invoicestatusid", insertable = false, updatable = false)
 	private InvoiceStatus invoiceStatus;
 	private Integer invoicestatusid;
 	
