@@ -2,13 +2,18 @@ package com.kindsonthegenius.fleetmsv2.fleet.models;
 
 import com.kindsonthegenius.fleetmsv2.parameters.models.Client;
 import com.kindsonthegenius.fleetmsv2.parameters.models.Location;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.data.domain.Auditable;
 import org.springframework.format.annotation.DateTimeFormat;
 
-import javax.persistence.*;
 import java.util.Date;
 
 @Entity
