@@ -1,10 +1,10 @@
 package com.kindsonthegenius.fleetmsv2.fleet.models;
 
 import com.kindsonthegenius.fleetmsv2.parameters.models.CommonObject;
+import jakarta.persistence.Entity;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
-import javax.persistence.Entity;
 
 @Entity
 @Data
