@@ -3,10 +3,10 @@ package com.kindsonthegenius.fleetmsv2.helpdesk.models;
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import com.kindsonthegenius.fleetmsv2.parameters.models.CommonObject;
+import jakarta.persistence.Entity;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
-import javax.persistence.Entity;
 
 @Entity
 @Data
